Factor out ActionBarActions
