+++ src/index.js
@@ -1,7 +1,7 @@
 export { ActionBar } from './ActionBar/ActionBar';
 export { ActionBarBack } from './ActionBar';
-export { ActionBarHeader } from './ActionBar/';
-export { ActionBarActions } from './ActionBar/ActionBar';
+export { ActionBarHeader } from './ActionBar';
+export { ActionBarActions } from './ActionBar';
 export { Alert } from './Alert/Alert';
 export { Badge } from './Badge/Badge';
 export { Label } from './Badge/Badge';
